+++ crates/types/src/field.rs
@@ -78,7 +78,7 @@ pub enum IdentityField {
 pub enum PrincipalField {
     Archive = ARCHIVE_FIELD,
     ParticipantIdentities = 45,
-    IdentityAddresses = 46,
+    IdentityAddresses = 110,
     DefaultCalendarId = 47,
     DefaultAddressBookId = 48,
     ActiveScriptId = 49,
@@ -157,7 +157,7 @@ impl From<PrincipalField> for u8 {
     fn from(value: PrincipalField) -> Self {
         match value {
             PrincipalField::ParticipantIdentities => 45,
-            PrincipalField::IdentityAddresses => 46,
+            PrincipalField::IdentityAddresses => 110,
             PrincipalField::DefaultCalendarId => 47,
             PrincipalField::DefaultAddressBookId => 48,
             PrincipalField::ActiveScriptId => 49,